Added benchmark for process execution (#100)
* Fixes process log race condition

Addresses a race condition where process logs could be truncated due to the process exiting before all logs were read.

This commit ensures that stdout and stderr are fully drained before the process is reaped, preventing the loss of buffered output, especially for fast-exiting commands. Also, codegen tests are skipped if codegen tools are not configured on the server.

* Added benchmark for process execution

* Remove codspeed token

* Force rerun

* Update version of go

* Upgrade version of codspeed

* Update benchmark

* Fix script

* Remove responseWritter

* Fix codegen configured

* Fix bench for write and delete file

* Rename test

* Update golang version